refactor: move PromptT into types package
