+++ python/mirascope/llm/calls/base_structured_context_call.py
@@ -6,9 +6,8 @@ from dataclasses import dataclass
 from typing import Generic
 
 from ..models import LLM
-from ..prompts import PromptT
 from ..tools import ContextToolDef
-from ..types import DepsT, FormatT, Jsonable, P
+from ..types import DepsT, FormatT, Jsonable, P, PromptT
 
 
 @dataclass
