+++ python/mirascope/llm/types/__init__.py
@@ -2,6 +2,7 @@
 
 from .dataclass import Dataclass
 from .jsonable import Jsonable
+from .prompts import PromptT
 from .type_vars import CovariantT, DepsT, FormatT, P, RequiredFormatT, ToolReturnT
 
 __all__ = [
@@ -11,6 +12,7 @@ __all__ = [
     "FormatT",
     "Jsonable",
     "P",
+    "PromptT",
     "RequiredFormatT",
     "ToolReturnT",
 ]
